+++ src/client/mod.rs
@@ -22,3 +22,5 @@ mod client_connection;
 
 pub use self::client::Client;
 pub use self::client_connection::ClientConnection;
+#[cfg(test)]
+pub use self::client_connection::test::TestClientConnection;