+++ src/rr/dnssec/trust_anchor.rs
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 use std::io::Cursor;
+use std::default::Default;
 
 use openssl::crypto::pkey::{PKey, Role};
 
@@ -21,12 +22,14 @@ use ::rr::dnssec::Algorithm;
 
 const ROOT_ANCHOR: &'static str = include_str!("Kjqmt7v.pem");
 
+// TODO: these should also store some information, or more specifically, metadata from the signed
+//  public certificate.
 pub struct TrustAnchor {
-  pkey: Vec<u8>
+  pkeys: Vec<Vec<u8>>
 }
 
-impl TrustAnchor {
-  pub fn new() -> TrustAnchor {
+impl Default for TrustAnchor {
+  fn default() -> TrustAnchor {
     let mut cursor = Cursor::new(ROOT_ANCHOR);
     let pkey = PKey::public_key_from_pem(&mut cursor).expect("Error parsing Kjqmt7v.pem");
     assert!(pkey.can(Role::Verify));
@@ -34,10 +37,23 @@ impl TrustAnchor {
 
     let alg = Algorithm::RSASHA256;
 
-    TrustAnchor{ pkey: alg.public_key_to_vec(&pkey) }
+    TrustAnchor{ pkeys: vec![alg.public_key_to_vec(&pkey)] }
+  }
+}
+
+impl TrustAnchor {
+  pub fn new() -> TrustAnchor {
+    TrustAnchor { pkeys: vec![] }
   }
 
   pub fn contains(&self, other_key: &[u8]) -> bool {
-    self.pkey == other_key
+    self.pkeys.iter().any(|k|other_key == k as &[u8])
+  }
+
+  /// inserts the trust_anchor to the trusted chain
+  pub fn insert_trust_anchor(&mut self, public_key: Vec<u8>) {
+    if !self.contains(&public_key) {
+      self.pkeys.push(public_key)
+    }
   }
 }